% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubR.R
\name{get_top_words_by_year}
\alias{get_top_words_by_year}
\title{Retrieve the Most Frequent Words in PubMed Titles for a Given Year}
\usage{
get_top_words_by_year(year, n_results = 1000, top_n = 15, pause = 1)
}
\arguments{
\item{year}{Numeric. The publication year (YYYY).}

\item{n_results}{Integer. Maximum number of articles to retrieve.}

\item{top_n}{Integer. Number of most frequent words to display.}

\item{pause}{Numeric. Time (in seconds) between batch API requests.}
}
\value{
A ggplot bar chart, invisibly returns dataframe of words and freq.
}
\description{
Searches PubMed articles from a specific year, extracts title words,
cleans them, removes stopwords and returns the most frequent ones.
}
