% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_data.R
\name{paste_data}
\alias{paste_data}
\title{Paste Data into R}
\usage{
paste_data(input = "clipboard", ...)
}
\arguments{
\item{input}{can be data from clipboard (default), a vector, data.frame, a
tibble or a (csv|tsv) file containing a data frame.}

\item{...}{other arguments passed to the functions
\code{\link[readr]{read_tsv}()} or \code{\link[readr]{read_csv}()}
depending on the input extension.}
}
\description{
Paste data from different sources (spreadsheet or data frame)
  into R.
}
