% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubR.R
\name{pubmed_top_authors}
\alias{pubmed_top_authors}
\title{Identify the Most Frequent Authors for a PubMed Keyword}
\usage{
pubmed_top_authors(keyword, max_results = 1000, batch_size = 300, top_n = 30)
}
\arguments{
\item{keyword}{Character. Search term.}

\item{max_results}{Integer. Maximum number of results to retrieve.}

\item{batch_size}{Integer. Batch size for summary retrieval.}

\item{top_n}{Integer. Number of top authors to display.}
}
\value{
A dataframe containing authors and frequencies (invisible).
}
\description{
Retrieves PubMed entries for a keyword, extracts authors from summaries,
counts occurrences, and plots the most frequent ones.
}
