% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubR.R
\name{pubmed_wordcloud}
\alias{pubmed_wordcloud}
\title{Generate a PubMed Word Cloud}
\usage{
pubmed_wordcloud(keyword, n_results = 300, pause = 1)
}
\arguments{
\item{keyword}{Character. Search term used in PubMed.}

\item{n_results}{Integer. Maximum number of results to retrieve.}

\item{pause}{Numeric. Delay (in seconds) between API requests to avoid rate limits.}
}
\value{
A dataframe of processed words and frequencies (invisible).
}
\description{
Searches PubMed for titles containing the selected keyword, cleans the text,
processes the most frequent terms and generates a word cloud.
}
