% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Custom stopword list used in PubR}
\format{
A character vector.
}
\usage{
stopwords
}
\description{
A character vector containing English stopwords used for text processing.
}
\keyword{datasets}
