% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubR.R
\name{top_journals}
\alias{top_journals}
\title{Most Common Journals for a PubMed Keyword}
\usage{
top_journals(keyword, max_results = 5000, top_n = 15, batch_size = 500)
}
\arguments{
\item{keyword}{Character. Search term.}

\item{max_results}{Integer. Maximum number of PubMed IDs to retrieve.}

\item{top_n}{Integer. Number of journals to display.}

\item{batch_size}{Integer. Number of summaries to fetch per batch.}
}
\value{
A ggplot bar chart showing the most frequent journals.
}
\description{
Retrieves PubMed entries using WebEnv history (avoiding URL overflow)
and returns a bar plot of the journals that publish most on the topic.
}
